+++ backend/Register/Domain/Repositories/IPortfolioRepository.cs
@@ -5,6 +5,6 @@ namespace backend.Register.Domain.Repositories;
 public interface IPortfolioRepository
 {
     Task<IEnumerable<Portfolio>> ListAsync();
-    Task<Portfolio> FindByIdAsync(long id);
+    Task<Portfolio> FindByIdAsync(int id);
     Task AddAsync(Portfolio portfolio);
 }
